2/22 Moving triangle!
Today I took a break from the opengl tutorials and figured out how to pass time into our vertex shader. I tried to do it with an in variable, but figured out that uniforms exist. so I made the time a uniform and got the triangle moving back in forth with a sine wave